Ported camera module to header-only
